requirements.txt
Django==4.2.6
mysqlclient==2.2.0
djangorestframework==3.14.0
python-decouple==3.8
django-filter==23.3
drf-yasg==1.21.7
coreapi==2.3.3
pypika==0.48.9
pytest-django==4.5.2
model_bakery==1.16.0